Changing UserRepo findById logic

+++ src/server/controllers/usersController.js
@@ -1,4 +1,6 @@
 
+const { validationResult } = require('express-validator');
+const formatErrors = require('../util/formatErrors.js');
 const RolesRepository = require('../repositories/RolesRepository.js');
 const UserRepository = require('../repositories/UserRepository.js');
 
@@ -9,23 +11,36 @@ const getUsers = async (req, res, next) => {
     const roles = await RolesRepository.getAllRoles();
     return res.status(200).json({ users, roles });
   } catch(error) {
-    return res.status(401).json({message: "There was a problem getting users"});
+    return res.status(400).json({message: "There was a problem getting users"});
   }
 };
 
 const deleteUser = async (req, res, next) => {
+
+	const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ message: formatErrors(errors.array())});
+  }
+
 	try {
-		console.log(req.body.userId)
-		// validate is integer
     await UserRepository.delete(req.body.userId);
     return res.status(200).json({ message: 'User successfully deleted' });
   } catch(error) {
-    return res.status(401).json({ message: "There was a problem deleting the user" });
+    return res.status(400).json({ message: "There was a problem deleting the user" });
   }
+};
+
+const patchUser = async (req, res, next) => {
+	const foundUser = await UserRepository.findById(req.params.userId);
+	console.log(foundUser)
+	console.log('Old Role Id: ' + foundUser.foundUser.role_id)
+	console.log('New Role Id: ' + req.body.roleId)
+	return res.status(200).json({ message: 'User successfully updated' });
 };
 
 
 module.exports = {
 	getUsers,
-	deleteUser
+	deleteUser,
+	patchUser
 }
